Improve custom Inputs
